requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aiohttp==3.9.3
aioprometheus==23.12.0
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argcomplete==1.10.3
astropy==6.1.0
astropy-iers-data==0.2024.6.3.0.31.14
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.8.2
bitsandbytes==0.43.1
blinker==1.7.0
blis==0.7.11
boilerpy3==1.0.7
boto3==1.34.124
botocore==1.34.124
cachetools==4.2.4
catalogue==2.0.10
certifi==2018.8.13
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
cmake==3.28.3
colorama==0.4.6
compressed-rtf==1.0.6
confection==0.1.4
contourpy==1.2.1
coverage==7.5.3
cryptography==42.0.7
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dask==2024.5.2
datasets==2.18.0
decorator==5.1.1
deepspeed==0.12.6
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
docutils==0.21.2
docx2txt==0.8
ebcdic==1.1.1
einops==0.7.0
elastic-transport==8.12.0
elasticsearch==8.11.1
elasticsearch-haystack==0.3.0
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
extract-msg==0.28.7
faiss-gpu==1.7.2
fastapi==0.110.2
filelock==3.13.1
fire==0.6.0
flash-attn==2.3.6
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-search-results==2.4.2
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
haystack-ai==2.0.0
haystack-bm25==1.0.2
hjson==3.1.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
hydra-core==1.3.2
idna==2.8
ijson==3.2.3
imagesize==1.4.1
IMAPClient==2.1.0
importlib_metadata==7.1.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.25.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jeepney==0.8.0
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==25.2.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy-imports==0.3.1
lightning-utilities==0.11.2
lit==18.1.1
lmdb==1.4.1
locket==1.0.0
lxml==5.2.2
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
mteb==1.2.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.2.1
nh3==0.2.17
ninja==1.11.1.1
nltk==3.8.1
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
openai==1.20.0
openpyxl==3.1.2
orjson==3.10.1
packaging==24.0
pandas==2.2.1
parso==0.8.4
partd==1.4.2
pathlib2==2.3.7.post1
pathspec==0.5.5
pdfminer.six==20191110
peft==0.7.1
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.11.0
pluggy==1.5.0
portalocker==2.8.2
posthog==3.5.0
preshed==3.0.9
prettyprint==0.1.5
prompt_toolkit==3.0.45
protobuf==3.20.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.13
pydantic_core==2.16.3
pyDeprecate==0.3.2
pyerfa==2.0.1.4
Pygments==2.18.0
pymongo==4.8.0
pynvml==11.5.0
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
pytest==8.2.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytrec-eval==0.5
pytz==2024.1
PyYAML==6.0.1
quantile-python==1.1
rank-bm25==0.2.2
ray==2.11.0
rdflib==7.0.0
readme_renderer==43.0
referencing==0.34.0
ReFinED @ https://github.com/amazon-science/ReFinED/archive/refs/tags/V1.zip#sha256=f70ef05f6dc644c758a919be65c960ba1f2cfb187f93d417962f3b79b3aeadc0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rouge==1.0.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
sacrebleu==2.4.2
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
SecretStorage==3.3.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
serpapi==0.1.5
shtab==1.7.1
six==1.12.0
smart-open==6.4.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SPARQLWrapper==2.0.0
SpeechRecognition==3.8.1
Sphinx==7.3.7
sphinx-automodapi==0.17.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.2.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.4.0
textract==1.6.5
thinc==8.2.3
threadpoolctl==3.3.0
tika==2.6.0
tiktoken==0.7.0
tokenizers==0.15.2
tomli==2.0.1
toolz==0.12.1
torch==2.1.2
torchmetrics==0.11.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.36.2
triton==2.1.0
trl==0.8.6
twine==5.1.0
typer==0.9.4
typing_extensions==4.10.0
tyro==0.8.4
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
Unidecode==1.3.8
urllib3==1.26.2
uvicorn==0.29.0
uvloop==0.19.0
vllm==0.2.6
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
websockets==12.0
Werkzeug==3.0.1
wget==3.2
xformers==0.0.23.post1
xlrd==1.2.0
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.1
